test 5

+++ src/controllers/nft.js
@@ -74,35 +74,38 @@ exports.uploadImageToIPFS = asyncHandler( async (req, res, next) => {
             fs.unlinkSync(nftImage.path);
             
         }
+
+        if( Array.isArray(req.files.custom) ) {
+
+            for(let customImage of req.files.custom) {
     
-        for(let customImage of req.files.custom) {
-    
-            const readableStreamForFile = fs.createReadStream(customImage.path);
-    
-            const options = {
-                pinataMetadata: {
-                    name: customImage.originalname,
-                    keyvalues: {
-                        customKey: 'customValue',
-                        issuedAt: Date.now().toString(),
-                        issuedBy: "naksh"
+                const readableStreamForFile = fs.createReadStream(customImage.path);
+        
+                const options = {
+                    pinataMetadata: {
+                        name: customImage.originalname,
+                        keyvalues: {
+                            customKey: 'customValue',
+                            issuedAt: Date.now().toString(),
+                            issuedBy: "naksh"
+                        }
+                    },
+                    pinataOptions: {
+                        cidVersion: 0
                     }
-                },
-                pinataOptions: {
-                    cidVersion: 0
+                };
+        
+                const upload = await pinata.pinFileToIPFS(readableStreamForFile, options);
+        
+                if(upload.IpfsHash == (undefined || null || "")) {
+                    return next(new ErrorResponse(400, "error uploading custom file to IPFS storage"));
                 }
-            };
-    
-            const upload = await pinata.pinFileToIPFS(readableStreamForFile, options);
-    
-            if(upload.IpfsHash == (undefined || null || "")) {
-                return next(new ErrorResponse(400, "error uploading custom file to IPFS storage"));
+        
+                custom.push(`${process.env.PINATA_PREVIEW_URL}${upload.IpfsHash}`);
+        
+                fs.unlinkSync(customImage.path);
             }
-    
-            custom.push(`${process.env.PINATA_PREVIEW_URL}${upload.IpfsHash}`);
-    
-            fs.unlinkSync(customImage.path);
-            
+
         }
         
         return res.status(200).send({ nftImageUrl, custom });
